+++ src/main/java/no/nav/pto/veilarbportefolje/aktiviteter/KafkaAktivitetMelding.java
@@ -13,7 +13,7 @@ public class KafkaAktivitetMelding {
     ZonedDateTime fraDato;
     ZonedDateTime tilDato;
     ZonedDateTime endretDato;
-    Integer version;
+    Long version;
     AktivitetTypeData aktivitetType;
     AktivitetStatus aktivitetStatus;
     boolean avtalt;
